refactor: buy warning message
