+++ webapp/src/components/BuyAssetPage/index.js
@@ -1,0 +1,3 @@
+import BuyWarningMessage from './BuyWarningMessage'
+
+export default BuyWarningMessage
